refactor(mcp): extract bridge action-log helpers
Audit finding L7 (`docs/reviews/crate-audits/convergio-mcp.md`):
`bridge.rs` mixed tool declarations, file logging, and integration-test
daemon bootstrapping in a 254-line file.

Move the file-backed action log (path resolution, log rotation, and
the HOME-resolution warn) into a sibling `bridge_log` module. The
bridge keeps its rmcp tool wiring and now exposes a thin
`Bridge::log_action` that just delegates. No behavior change.

Tracks: T6e0da979-90e7-473f-a2e6-2f10f21bfc3f
