+++ crates/convergio-mcp/src/bridge.rs
@@ -1,5 +1,6 @@
 //! MCP tool declarations for the bridge.
 
+use crate::bridge_log;
 use crate::help;
 use crate::http::fallback_error;
 use convergio_api::{ActRequest, Action, HelpRequest, HelpTopic, HelpVerbosity};
@@ -10,11 +11,7 @@ use rmcp::{
 };
 use schemars::JsonSchema;
 use serde_json::Value;
-use std::fs::{self, OpenOptions};
-use std::io::Write;
-use std::path::PathBuf;
 use std::sync::Arc;
-use std::time::{SystemTime, UNIX_EPOCH};
 use tokio::sync::Mutex;
 
 #[derive(Clone)]
@@ -80,53 +77,7 @@ impl Bridge {
     }
 
     pub(crate) fn log_action(&self, action: Action, response: &convergio_api::AgentResponse) {
-        let path = match mcp_log_path() {
-            Ok(p) => p,
-            Err(e) => {
-                // Surface the setup failure through tracing so MCP
-                // diagnostics aren't silently dropped when HOME is
-                // unset/non-unicode. VarError carries no payload data.
-                tracing::warn!(error = %e, "mcp log path unavailable; skipping action log");
-                return;
-            }
-        };
-        if let Some(parent) = path.parent() {
-            let _ = fs::create_dir_all(parent);
-        }
-        trim_log(&path);
-        let ts = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|d| d.as_secs())
-            .unwrap_or_default();
-        let line = serde_json::json!({
-            "ts": ts,
-            "action": action.as_str(),
-            "ok": response.ok,
-            "code": response.code,
-            "next": response.next,
-        });
-        if let Ok(mut file) = OpenOptions::new().create(true).append(true).open(&path) {
-            let _ = writeln!(file, "{line}");
-        }
-    }
-}
-
-fn mcp_log_path() -> Result<PathBuf, std::env::VarError> {
-    Ok(PathBuf::from(std::env::var("HOME")?).join(".convergio/mcp.log"))
-}
-
-fn trim_log(path: &std::path::Path) {
-    const MAX_BYTES: u64 = 256 * 1024;
-    let Ok(meta) = fs::metadata(path) else {
-        return;
-    };
-    if meta.len() <= MAX_BYTES {
-        return;
-    }
-    if let Ok(content) = fs::read_to_string(path) {
-        let keep: Vec<&str> = content.lines().rev().take(1000).collect();
-        let trimmed = keep.into_iter().rev().collect::<Vec<_>>().join("\n");
-        let _ = fs::write(path, format!("{trimmed}\n"));
+        bridge_log::append(action, response);
     }
 }
